The AI now considers diagonal moves. This completes the main AI algorithm (we can tweak it a bit to yield desired results, but overall this should be good).
Informal tests show that it is not bad. I accidentally let it beat me and it
tied Jeff Crawford. Overall, the AI is extremely defensive, so it is _probably_
possible to corner it into making a bad move, but it is also a bit unpredictable,
so doing so is not easy.